Clear changes after reading
This is an amendment to commit a73dd28950f35958503b38c94979ded5872cebf1
When Worksheets are added to a Workbook in #add_worksheet, this is
recorded as a change. Since Excel::Reader uses this method (and may in
the future use other methods that record changes), it is necessary to
clear all changes after the reading process.