Added bin/xlsopcodes
xlsopcodes takes as arguments the path to an .xls-file which is to be
analyzed, and optionally the path to an output file (default is STDOUT).
It will then print in order all opcodes (hex-value and name), the length
of the attached data and the data itself.
This tool facilitates tracking down bugs, both in the reader and the
writer.