Speed up writing: More efficient #format_dates!
This method was very liberal with the number of new formats it created.
Roughly, it created a new format for each Date/Time cell it detected.

In a large sheet with many date/time cells, the high number of
generated formats meant much slower writing, specifically when it the
writers enter the #write_xfs and #write_rows methods.

This commit makes format_dates! less liberal with generating new
formats.  Essentially, when it needs to generate a new format to assign
to a date/time cell, it will re-use an existing format that matches the
needed characteristics.